+++ lib/simple_form/bootstrap/form_builders/token.rb
@@ -1,0 +1,3 @@
+SimpleForm::FormBuilder.class_eval do
+  map_type :token, to: SimpleForm::Bootstrap::Inputs::TokenInput
+end