+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 ### enhancements
+ * Implement support for token fields. Use the `as: :token` on a Simple Form input field.
 
 ### bug fixes
  * Do not change selects to the Bootstrap select class unless they are explicitly select fields.